fix: data type to allocate correct memory byte.

+++ adc_calculation.c
@@ -45,7 +45,7 @@ ISR(TIMER1_COMPA_vect){
 	unsigned int raw = ADC; // 16bit
 	U[0] = raw>>2; // 10bit to 8bit
 	int8_t *B = PINC&(1<<7) ? &T : &S; //
-	int result = 0; // 20bit
+	long result = 0; // 20bit
 	for (int i=0;i<16;i++){
 		result += U[i] * (B[i]);
 	}
